:recycle: refactor test code
