--- example/.babelrc
@@ -1,10 +1,0 @@
-{
-  "plugins": [
-    [
-      "../dist/index.js",
-      {
-        "minJSONStringSize": 0
-      }
-    ]
-  ]
-}
