+++ package.json
@@ -7,7 +7,7 @@
     "example": "babel example/input.js -o example/output.js && node example/output.js",
     "fmt": "prettier --write \"src/**/*.ts\"",
     "lint": "eslint 'src/**/*.ts' --fix",
-    "test": "jest"
+    "test": "ls test/__fixtures__/**/*_out.js | xargs -L1 node && jest"
   },
   "dependencies": {
     "@babel/core": "^7.7.7",
