+++ test/__fixtures__/array/array.js
@@ -1,0 +1,2 @@
+const var1 = [1, "two", {three: 3}]
+const var2 = [{one: 1}, {two: 2}, {three: 3}]
