+++ test/__fixtures__/array/array_out.js
@@ -1,0 +1,2 @@
+const var1 = JSON.parse('[1,"two",{"three":3}]')
+const var2 = JSON.parse('[{"one":1},{"two":2},{"three":3}]')
