+++ test/__fixtures__/object/option.js
@@ -1,0 +1,1 @@
+const var1 = { b: 1, c: 2 }
