+++ test/__fixtures__/object/option_out.js
@@ -1,0 +1,4 @@
+const var1 = {
+  b: 1,
+  c: 2
+}
