+++ test/__fixtures__/object/valid_case.js
@@ -1,0 +1,15 @@
+const var1 = { foo: 'foo' }
+const var2 = { foo: "foo" }
+const var3 = { foo: "fo'o" }
+const var4 = { foo: "fo\'o" }
+const var5 = { foo: "fo\"o" }
+const var6 = { foo: 'fo"o' }
+const var7 = { foo: 'fo\"o' }
+const var8 = { foo: 'fo\'o' }
+const var9 = { foo: 10 }
+const var10 = { foo: true }
+const var11 = { foo: null }
+const var12 = { foo: [null, 10, 'foo'] }
+const var13 = { foo: [null, [10, 2], [{ foo: 'foo' }]] }
+const var14 = { foo: { bar: 1337 } }
+const var15 = { 1: "123", 23: 45, b: "b_val" }
