+++ test/__fixtures__/object/valid_case_out.js
@@ -1,0 +1,15 @@
+const var1 = JSON.parse('{"foo":"foo"}')
+const var2 = JSON.parse('{"foo":"foo"}')
+const var3 = JSON.parse('{"foo":"fo\'o"}')
+const var4 = JSON.parse('{"foo":"fo\'o"}')
+const var5 = JSON.parse('{"foo":"fo\\"o"}')
+const var6 = JSON.parse('{"foo":"fo\\"o"}')
+const var7 = JSON.parse('{"foo":"fo\\"o"}')
+const var8 = JSON.parse('{"foo":"fo\'o"}')
+const var9 = JSON.parse('{"foo":10}')
+const var10 = JSON.parse('{"foo":true}')
+const var11 = JSON.parse('{"foo":null}')
+const var12 = JSON.parse('{"foo":[null,10,"foo"]}')
+const var13 = JSON.parse('{"foo":[null,[10,2],[{"foo":"foo"}]]}')
+const var14 = JSON.parse('{"foo":{"bar":1337}}')
+const var15 = JSON.parse('{"1":"123","23":45,"b":"b_val"}')
